third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-reset-value-after-reload-expected.txt

Checks reloading page resets input values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testInput1.value is ""
PASS testInput2.value is "02:02"
PASS successfullyParsed is true

TEST COMPLETE
